backend/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-vote==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
gunicorn==20.1.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
